audit-remote: workaround for BZ#1534748
In RHEL-7.5 audit was updated to listen on both IPv4 and IPv6 when
remote logging is used. However, due to bug BZ#1534748 addr field
in audit records is incorrectly handled and set to 0.0.0.0 when
IPv6 is used. Therefore we need temporarily enforce using IPv4 by
disabling IPv6 networking on TOE.

Signed-off-by: Ondrej Moris <omoris@redhat.com>